backend/requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# NLP和ML
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0
python-dateutil>=2.8.2

# Google Calendar API
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Outlook Calendar API
msal>=1.25.0
requests>=2.31.0

# 数据处理
pandas>=2.1.0
pydantic>=2.5.0

# 开发工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.0.0

# 其他
cryptography>=41.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
